backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sentence-transformers
numpy
PyPDF2
python-docx
httpx
pydantic
